Working EL7 Genesis
This needs to be cleaned up a bit but is the basis for a working
genesis on EL7. For whatever reason network-prep.init worked properly
before however it no longer works when converting from sysv to systemd.
I believe what is happening is the script runs and gets the name of eth0
but that gets renamed to ens3 later during init and causes dhcp to never
start.